requirements/base.txt
##########
#  BASE  #
##########


# Bleeding edge Django
django==1.10.5
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
django-filter==1.0.1
markdown==2.6.7
coreapi==2.1.1
django-cors-headers==1.2.0

# Configuration
django-environ==0.4.1
django-secure==1.0.1

# Permissions
django-guardian==1.4.6

# Models
django-model-utils==2.6.0

# For the persistence stores
psycopg2==2.6.2

# Time zones support
pytz==2016.10

# HTTP Requests
requests==2.13.0
requests-oauthlib==0.8.0
